+++ apps/io-session-manager-oi/src/adapters/outbound/config-loader.ts
@@ -3,7 +3,10 @@ import {
   type ConfigLoader,
 } from "@pagopa/io-env-config";
 
-import { ConfigSchema, type Config } from "../../domain/entities/config.js";
+import {
+  ConfigSchema,
+  type Config,
+} from "../../domain/entities/config.entity.js";
 
 export const createConfigLoader = (): ConfigLoader<Config> =>
   createEnvConfigLoader(ConfigSchema);
