+++ apps/io-session-manager-oi/src/application/use-cases/load-config.use-case.ts
@@ -1,7 +1,7 @@
 import type { UseCase } from "@pagopa/io-core-domain";
 import type { ConfigError, ConfigLoader } from "@pagopa/io-env-config";
 
-import type { Config } from "../../domain/entities/config.js";
+import type { Config } from "../../domain/entities/config.entity.js";
 
 export const loadConfigUseCase =
   (
