+++ apps/io-session-manager-oi/src/domain/entities/config.entity.ts
